Move parameter parsing to LogicalPlanBuilder
